Issue #1277: Initial work on setting precision, doesn't cut off # (1.2345678901234 -> 1.23500...1)
This code does not accurately set decimal places properly, but it sets the number of digits that the number has (e.g. 1234.5 -> 1234), and
it does not reduce the filesize, because it keeps the rounded version and doesn't cut them off after 4 decimal places (e.g. 934.000000000...1
is kept or 8.379999999999...9 is kept). Need to look into the double datatype it is using and potentially change the datatype that is used.
Also need to update helper method Util::setPrecision() so it properly sets a number to have 4 *decimal places*, not digits. Tested on Arch Linux.